feat: add findAll method to PackageRepository
This commit adds the `findAll` method to the `PackageRepository` abstract class. This method returns a promise that resolves to an array of `Package` objects. The implementation of this method is added to the `InMemoryPackageRepository` class.

Refactor on inMemoryDataBase
test: enhance MarkAsDelivered use case tests and improve logging in repositories
feat: add findByDeliveryManId method and implement GetPackagesByDeliveryMan use case
feat: basic setup